# src/app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {

    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "mmonitor.gost-group.com",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/dashboard/admin/layers": {
      "get": {
        "tags": [
          "layer-administration-controller"
        ],
        "summary": "Запрос на получение слоёв с фильтрацией и пагинацией.",
        "description": "Метод позволяет получить слои.",
        "operationId": "searchLayersUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "description": "enabled",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "field",
            "in": "query",
            "description": "field",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "paged",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "sort.sorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "sort.unsorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "unpaged",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«PageResult«LayerDTO»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/admin/layers/edit": {
      "post": {
        "tags": [
          "layer-administration-controller"
        ],
        "summary": "Запрос редактирование полей слоя",
        "description": "Метод позволяет редактировать информацию слоя.",
        "operationId": "editLayerUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "info",
            "description": "info",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LayerDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/admin/layers/enable": {
      "post": {
        "tags": [
          "layer-administration-controller"
        ],
        "summary": "Запрос позволяет включать и выключать слой",
        "operationId": "editEnableStatusUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "layerAlias",
            "description": "layerAlias",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LayerDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/admin/layers/layer_type": {
      "get": {
        "tags": [
          "layer-administration-controller"
        ],
        "summary": "getLayerTypes",
        "operationId": "getLayerTypesUsingGET",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«Map«string,string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/dashboard/legends": {
      "get": {
        "tags": [
          "legend-controller"
        ],
        "summary": "Легенда цветового заполнения карты",
        "description": "Метод выводит легенду цветового заполнения карты",
        "operationId": "legendsUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "paged",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "sort.sorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "sort.unsorted",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "unpaged",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«PageResult«LegendEditingDTO»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/legends/level_type": {
      "get": {
        "tags": [
          "legend-controller"
        ],
        "summary": "Получение описания к полю level",
        "description": "Метод позволяет получить описание к полю level",
        "operationId": "getLayerTypesUsingGET_1",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«Map«string,string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/legends/save": {
      "post": {
        "tags": [
          "legend-controller"
        ],
        "summary": "Метод добавления/изменения легенды",
        "description": "Метод добавляет/изменяет легенду",
        "operationId": "saveLegendUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "legend",
            "description": "Легенда",
            "required": false,
            "schema": {
              "$ref": "#/definitions/LegendEditingDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«LegendEditingDTO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/legends/season_type": {
      "get": {
        "tags": [
          "legend-controller"
        ],
        "summary": "Получение описания к полю season",
        "description": "Метод позволяет получить описание к полю season",
        "operationId": "getSeasonTypesUsingGET",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«Map«string,string»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/dashboard/legends/{id}": {
      "delete": {
        "tags": [
          "legend-controller"
        ],
        "summary": "Метод удаления легенды по ID",
        "description": "Метод позволяет удалить легенду по её идентификатору",
        "operationId": "deleteLegendUsingDELETE",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор легенды",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/dashboard/property": {
      "get": {
        "tags": [
          "property-controller"
        ],
        "summary": "Получение настроек",
        "description": "Метод позволяет получить настройки для фронта",
        "operationId": "getPropertiesUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "types",
            "in": "query",
            "description": "types",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CAMERA",
                "DASHBOARD"
              ]
            },
            "collectionFormat": "multi",
            "enum": [
              "CAMERA",
              "DASHBOARD"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«string»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "property-controller"
        ],
        "summary": "Изменение настроек",
        "description": "Метод позволяет изменить значение настроек",
        "operationId": "savePropertyValueUsingPATCH",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "property",
            "description": "Измененная настройка",
            "required": false,
            "schema": {
              "$ref": "#/definitions/KeyPropertyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RestResult«string»"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "KeyPropertyDTO": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "description": "Описание настройки"
        },
        "enabled": {
          "type": "boolean",
          "description": "Состояние настроки - ВКЛ/ВЫКЛ"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Идентификатор настройки"
        },
        "key": {
          "type": "string",
          "description": "Ключ"
        },
        "type": {
          "type": "string",
          "description": "Тип",
          "enum": [
            "CAMERA",
            "DASHBOARD"
          ]
        },
        "value": {
          "type": "string",
          "description": "Значение настройки"
        }
      },
      "title": "KeyPropertyDTO",
      "description": "Настройки"
    },
...
    "LayerDTO": {
      "type": "object",
      "properties": {
        "alias": {
          "type": "string",
          "description": "Алиас слоя"
        },
        "businessType": {
          "type": "string",
          "description": "Тип слоя",
          "enum": [
            "alias",
            "dimension",
            "UNKNOWN"
          ]
        },
        "creationDate": {
          "type": "string",
          "format": "date-time",
          "description": "Дата создания"
        },
        "description": {
          "type": "string",
          "description": "Описание слоя"
        },
        "enabled": {
          "type": "boolean",
          "description": "Активирован ли слой"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Идентификатор слоя"
        },
        "modificationDate": {
          "type": "string",
          "format": "date-time",
          "description": "Дата модифицирования"
        },
        "name": {
          "type": "string",
          "description": "Имя слоя"
        },
        "revisionId": {
          "type": "integer",
          "format": "int64",
          "description": "Ревизия данных"
        },
        "sortOrder": {
          "type": "integer",
          "format": "int32",
          "description": "Порядок сортировки"
        },
        "srcType": {
          "type": "string",
          "description": "Тип слоя"
        }
      },
      "title": "LayerDTO",
      "description": "Сущность слоя."
    },
...
    "LegendEditingDTO": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string",
          "description": "Отображаемый цвет"
        },
        "colorValue": {
          "type": "string",
          "description": "Значение цвета"
        },
        "description": {
          "type": "string",
          "description": "Легенда"
        },
        "emoji": {
          "type": "string",
          "description": "Смайлик"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Идентификатор легенды"
        },
        "level": {
          "type": "string",
          "description": "Уровень отображения",
          "enum": [
            "moscow.russia",
            "regions.moscow.russia",
            "districts.regions.moscow.russia",
            "report",
            "dashboard",
            "dashboard.dashlet",
            "ts",
            "UNKNOWN"
          ]
        },
        "levelDescription": {
          "type": "string",
          "description": "Описание уровня"
        },
        "maxPercent": {
          "type": "integer",
          "format": "int32",
          "description": "Максимальный процент по записи"
        },
        "minPercent": {
          "type": "integer",
          "format": "int32",
          "description": "Минимальный процент по записи"
        },
        "report": {
          "type": "boolean",
          "description": "Флаг использования цветов в Диаграммах отчёта"
        },
        "season": {
          "type": "string",
          "description": "Сезон",
          "enum": [
            "all",
            "winter",
            "SYS",
            "svod"
          ]
        },
        "seasonDescription": {
          "type": "string",
          "description": "Описание сезона"
        },
        "sortOrder": {
          "type": "integer",
          "format": "int32",
          "description": "Порядок сортировки"
        },
        "theme": {
          "type": "string",
          "description": "Тема"
        }
      },
      "title": "LegendEditingDTO",
      "description": "Легенда для карты, редактируемая"
    },
...
    "PageResult«LayerDTO»": {
      "type": "object",
      "properties": {
        "objects": {
          "type": "array",
          "description": "Обекты страницы",
          "items": {
            "$ref": "#/definitions/LayerDTO"
          }
        },
        "pageNumber": {
          "type": "integer",
          "format": "int32",
          "description": "Номер страницы"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32",
          "description": "Количество элементов на странице"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64",
          "description": "Общее количество элементов"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32",
          "description": "Общее количество страниц"
        }
      },
      "title": "PageResult«LayerDTO»"
    },
    "PageResult«LegendEditingDTO»": {
      "type": "object",
      "properties": {
        "objects": {
          "type": "array",
          "description": "Обекты страницы",
          "items": {
            "$ref": "#/definitions/LegendEditingDTO"
          }
        },
        "pageNumber": {
          "type": "integer",
          "format": "int32",
          "description": "Номер страницы"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32",
          "description": "Количество элементов на странице"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64",
          "description": "Общее количество элементов"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32",
          "description": "Общее количество страниц"
        }
      },
      "title": "PageResult«LegendEditingDTO»"
    },
...
    "RestResult": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "type": "object"
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult"
    },
...
    "RestResult«LegendEditingDTO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "$ref": "#/definitions/LegendEditingDTO"
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult«LegendEditingDTO»"
    },
...
    "RestResult«Map«string,string»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult«Map«string,string»»"
    },
...
    "RestResult«PageResult«LayerDTO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "$ref": "#/definitions/PageResult«LayerDTO»"
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult«PageResult«LayerDTO»»"
    },
    "RestResult«PageResult«LegendEditingDTO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "$ref": "#/definitions/PageResult«LegendEditingDTO»"
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult«PageResult«LegendEditingDTO»»"
    },
...
    "RestResult«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "enum": [
            "SUCCESS",
            "ERROR"
          ]
        },
        "data": {
          "type": "string"
        },
        "errorCode": {
          "type": "integer",
          "format": "int32"
        },
        "requestId": {
          "type": "string"
        },
        "responseId": {
          "type": "string"
        },
        "stackTrace": {
          "type": "string"
        },
        "synopsis": {
          "type": "string"
        }
      },
      "title": "RestResult«string»"
    },
...
  }
}